Add actual completion to Combinator bash support
Previously, the bash completion would only display names to be completed, it
wouldn't complete them for you.  This fixes a number of subtle interactions
with implementation details of the bash completion library (like the
requirement for a "local" variable named 'cur' for _filedir to work), as well
as making use of 'compgen' where required instead of simply setting COMPREPLY.

Also re-indent the file according to Python-like guidelines.

This also strips out the incredibly annoying and ever-present '/' on the end of
completed branch names.

Author: glyph
Reviewer: exarkun