Allow Combinator's "mkbranch" command to work with versions later than 1.4.
Sorry for the long delay on merging this; without any tests in combinator I tend to be pretty cagey about making changes.  I've now used this on a bunch of different machines for several days and I'm more or less comfortable that it works, but we should fix the test issue soon.

Author: wsanchez

Reviewer: glyph

Fixes #1536