Merge safer-mkbranch-2384-2
Author: exarkun
Reviewer: glyph, washort
Fixes #2384

Add a check to mkbranch so that it refuses to try to create a new branch with
the same name as an existing branch.  Previously this would put a directory
named trunk into the existing branch.  Now it will exit early without changing
the repository or any working copies.